feat(quorum): P1-3 Quorum 网络集成 + 测试覆盖率优化 (#62)
* docs(spike): 重命名 Layer 1 完整性分析报告，添加日期前缀

- 重命名文件：layer1-completeness-analysis.md -> 2026-02-13_layer1-completeness-analysis.md
- 遵循 spike 文档命名规范（YYYY-MM-DD_主题.md）

* docs(pm): 添加 Layer 1 一致性层 P0+P1 优先级改进 Pre 文档

- 创建 PR 全流程文档：2026-02-13_PR-layer1-consistency-improvements_全流程.md
- 包含 P0 优先级改进：2PC Gossip 状态同步、事务状态持久化
- 包含 P1 优先级改进：超时处理优化、状态转换验证、Quorum 网络集成
- 工作量评估：10 个工作日
- 待架构师评审后启动开发

* docs(spike): 批量重命名 spike 文档，统一添加日期前缀

- 重命名 18 个 spike 文档，遵循 YYYY-MM-DD_主题.md 命名规范
- 涉及文档：
  - kv-storage-engine-arch-analysis.md (2026-02-09)
  - pending-tasks-plan.md (2026-02-10)
  - bftree-research.md, consistency-layering-research.md 等 (2026-02-11)
  - DOUBAO-Metadata-Merkle-Tree元数据版本控制.md 等 (2026-02-12)
  - checkpoint_2026-01-19_wal-snapshot-checkpoint-design.md 调整为 2026-01-19_checkpoint_wal-snapshot-checkpoint-design.md
- 现在 100% spike 文档都遵循统一命名规范

* docs(pm): 加强压力测试要求，提升测试强度 5-10 倍

变更内容：
- 压力测试从「可选」改为「✅ 必要」
- 明确仅在本地运行，不在 GitHub CI 上运行
- 并发事务处理：100 → 1000 并发（10倍）
- Gossip 消息吞吐：1000 → 10,000 msg/s（10倍）
- 持久化性能：1000 → 5000 tx/s（5倍）
- 新增故障恢复压力测试：100 次故障模拟
- 新增内存压力测试：100 万事务，< 2GB
- 新增并发竞态检测：go test -race
- 添加压力测试运行方式说明（make test-stress）

* docs(pm): 根据架构师评审意见修复 Pre 文档关键问题

根据架构师第1轮评审意见，修复 3 个关键问题：

1. 命名空间冲突（🔴 高优先级）
   - 修复前：const NamespaceTx = "tx:"
   - 修复后：const NamespaceTx = "meta:tx:"
   - 原因：保持与现有命名空间一致性

2. 字段名错误（🔴 高优先级）
   - 修复前：tx.ID
   - 修复后：tx.TxID
   - 原因：与实际代码结构体字段名一致

3. 压力测试指标过高（🟡 中优先级）
   - 并发事务：1000 → 500（单机合理负载）
   - Gossip 吞吐：10,000 → 5,000 msg/s（参考 Raft 上限）
   - 持久化性能：5,000 → 1,000 tx/s（MVStore 实际性能）

其他改进：
- 工作量评估：10 天 → 16 天
- 添加架构师第1轮评审记录
- 添加架构师强烈建议的 3 个优化项
- 修正存储接口：c.mvstore.Put() → c.metadataKV.PutRaw()

待第2轮架构师评审通过后可启动开发。

* docs(pm): 将架构师强烈建议改为必须实施项

变更内容：
1. 将「架构师强烈建议（可选优化项）」改为「架构师必须实施项（必选）」
2. 3 个强制优化项：
   - 使用 sync.Map 替代 map（并发安全优化）
   - 实现批量写入缓冲区（持久化性能优化）
   - 实现 Gossip 消息限流（避免消息风暴）
3. 工作量评估：16 天 → 18 天（新增 2 天强制优化项）
4. 功能目标新增：强制优化项
5. 更新架构师第1轮评审记录

预计完成日期：2026-03-05（18 个工作日）

* docs(pm): 修复第2轮架构师评审发现的 2 处代码错误

根据第2轮架构师评审意见，修复 2 处遗留代码错误：

1. recoverTransactions() 使用错误的存储字段（第187行）
   - 修复前：c.mvstore.Scan(NamespaceTx, "")
   - 修复后：c.metadataKV.ScanRaw(context.Background(), NamespaceTx, "")

2. recoverTransactions() 使用错误的字段名和存储方式（第192行）
   - 修复前：c.transactions[tx.ID] = tx
   - 修复后：c.transactions.Store(tx.TxID, tx)
   - 说明：使用 sync.Map 方式（符合强制优化项 7.1）和正确的字段名

同时更新：
- 架构师评审记录：新增第2轮评审结果
- 预审批确认：更新为待第3轮评审

所有代码示例现已符合实际代码结构和强制优化项要求。

* docs(pm): 第3轮架构师评审通过，更新评审记录和预审批确认

第3轮架构师评审结果：
- ✅ 评审通过
- ✅ 所有 7 个修复点已正确更新
- ✅ 代码示例与实际代码结构完全一致
- ✅ 强制优化项设计合理
- ✅ 风险评估全面

更新内容：
1. 评审记录：新增第3轮评审结果（✅ 通过）
2. 预审批确认：更新为架构师签字通过，同意开工

启动开发许可：✅ 同意开工
架构师签字：👤 架构师 2026-02-13

Pre 文档已完成所有评审，可以启动开发。

* feat(transport): 添加 TwoPCGossipPayload 消息类型（P0-1 第1步）

P0-1: 2PC Gossip 状态同步 - 第1步完成

变更内容：
1. 新增 MessageTypeTwoPCGossip = 10
2. 新增 TwoPCGossipPayload 结构体
   - Phase 字段区分三种消息类型：state/query/reply
   - TxID: 事务ID
   - State: 事务状态
   - Coordinator: 协调者地址
   - Timestamp: HLC 时间戳
   - Success: 查询是否成功（仅用于 reply）
   - Requester: 查询发起者
   - MessageID: 消息去重ID
3. 更新 String() 方法：新增 "TWOPC_GOSSIP"
4. 更新 IsValid() 方法：最大 MessageType 检查
5. 添加到 payloadTypeFactories 映射
6. 更新 EncodePayload 方法：支持 *TwoPCGossipPayload

下一步：在 twopc_coordinator.go 中实现 Gossip 状态扩散和查询方法

* feat(consistency): 实现 2PC Gossip 状态同步方法（P0-1 第2步）

P0-1: 2PC Gossip 状态同步 - 第2步完成

变更内容：
1. 修改 TwoPCTransaction 结构体
   - 新增 Coordinator 字段：协调者节点 ID
   - 新增 Acknowledgments 字段：响应追踪

2. 更新 NewTwoPCTransaction 函数
   - 初始化 Acknowledgments 字段

3. 新增 Gossip 状态同步方法
   - gossipTransactionStates(): 周期性 Gossip 状态扩散（每 5 秒）
   - queryTransactionDecision(): Gossip 查询事务决策
   - handleGossipStateMessage(): 处理 Gossip 状态消息
   - handleGossipQueryMessage(): 处理 Gossip 查询消息
   - handleGossipReplyMessage(): 处理 Gossip 响应消息

实现要点：
- 使用 transport.TwoPCGossipPayload 统一消息格式
- Phase 字段区分 state/query/reply 三种消息类型
- 使用 MessagePackCodec 进行消息编解码
- 异步发送 Gossip 状态消息，避免阻塞
- TODO: 集成集群管理器获取节点列表
- TODO: 实现 HLC 时间戳比较（P1-2）

下一步：编写单元测试和集成测试

* feat(consistency): 集成 Gossip 状态同步任务和消息处理器（P0-1 第3-4步）

P0-1: 2PC Gossip 状态同步 - 第3-4步完成

变更内容：
1. 修改 TwoPCMerkleCoordinator 结构体
   - 新增 gossipInterval: Gossip 间隔（默认 5 秒）
   - 新增 gossipStopChan: 停止信号
   - 新增 gossipWaitGroup: 等待 goroutine 退出

2. 新增统一消息处理器
   - handleGossipMessage(): 根据 Phase 字段分发消息
   - 支持 state/query/reply 三种消息类型

3. 新增 Gossip 任务管理方法
   - startGossipTask(): 启动周期性 Gossip 任务
   - stopGossipTask(): 优雅停止 Gossip 任务

4. 更新构造函数
   - NewTwoPCMerkleCoordinator: 初始化 Gossip 字段
   - NewTwoPCMerkleCoordinatorWithTransport: 启动 Gossip 任务
   - 注册 Gossip 消息处理器（TODO: 实现消息分发器）

5. 更新事务管理
   - BeginTransaction: 设置协调者信息（localNodeID）
   - Close: 停止 Gossip 任务

实现要点：
- 使用 ticker 实现周期性任务（5 秒间隔）
- 使用 stopChan 实现 goroutine 优雅退出
- 使用 WaitGroup 等待 goroutine 完成
- TODO: 实现统一的消息分发器，支持多个处理器注册

下一步：编写单元测试和集成测试

* fix(consistency): 修复 2PC Gossip 消息解码逻辑（P0-1 Step 5）

问题：
- handleGossipStateMessage/QueryMessage/ReplyMessage 使用 codec.DecodeFromBytes()
- 但 Transport 传递的是 msg.Payload（纯 MessagePack 数据，不含 TLV 头）
- 导致解码时将 payload 前几个字节误读为 Type+Length，产生"消息过大"错误

解决方案：
- 添加 msgpack/v5 导入
- 三个 handler 方法直接使用 msgpack.Unmarshal() 解码 payload
- handleGossipMessage 分发器同样修复

测试结果：
- 7 个 Gossip 相关测试全部通过
- consistency 和 transport 包测试全部通过

* feat(consistency): 实现事务状态持久化（P0-2）

新增功能：
- 添加 NamespaceTx 命名空间（meta:tx:）
- 实现 persistTransaction() - 持久化事务到 MetadataKV
- 实现 loadTransaction() - 从 MetadataKV 加载事务
- 实现 recoverTransactions() - 启动时恢复未完成事务
- 实现 deleteTransaction() - 删除已完成事务记录

集成点：
- NewTwoPCMerkleCoordinatorWithTransport() 启动时调用 recoverTransactions()
- PreCommit/Commit/Rollback 时调用 persistTransaction()
- 只恢复 PreCommit 状态的事务，跳过终态事务

单元测试：
- TestPersistTransaction - 测试事务持久化
- TestLoadTransaction - 测试事务加载
- TestRecoverTransactions - 测试事务恢复（含状态过滤）

容错设计：
- 持久化失败：记录日志，继续内存操作
- 加载失败：跳过该事务，不影响其他恢复
- 序列化失败：记录错误，跳过该事务

* refactor(consistency): 使用 sync.Map 替代 map（强制优化 7.1）

变更内容：
- TwoPCMerkleCoordinator.transactions: map[string]*TwoPCTransaction -> sync.Map
- 移除所有 c.mu 锁对 transactions 的保护（sync.Map 自带并发安全）

更新的方法：
- BeginTransaction: c.transactions.Store(txID, tx)
- GetTransaction: c.transactions.Load(txID)
- GetAllTransactions: c.transactions.Range()
- CleanupTimeoutTransactions: c.transactions.Range()
- gossipTransactionStates: c.transactions.Range()
- queryTransactionDecision: c.transactions.Load(txID)
- handlePreCommitResponse: c.transactions.Load(txID)
- handleGossipStateMessage: c.transactions.Load(txID)
- handleGossipQueryMessage: c.transactions.Load(txID)
- handleGossipReplyMessage: c.transactions.Load(txID)
- recoverTransactions: c.transactions.Store(txID, tx)
- Close: c.transactions.Range()
- Commit: c.transactions.Delete(txID)
- rollbackTransaction: c.transactions.Delete(txID)

性能优化：
- 读多写少场景下 sync.Map 比 map + RWMutex 性能更好
- 减少全局锁竞争，提高并发性能

* feat(consistency): 实现事务持久化批量写入缓冲区（强制优化 7.2）

新增文件：
- tx_persistence_buffer.go: TransactionPersistenceBuffer 实现
- tx_persistence_buffer_test.go: 缓冲区单元测试

核心功能：
- 批量写入：积累多个事务后一次性写入，减少 IO 次数
- 定时刷新：100ms 间隔，避免延迟过高
- 批量阈值：100 个事务触发立即刷新
- 并发安全：mutex 保护缓冲区
- 优雅关闭：Close() 时刷新剩余事务

集成点：
- TwoPCMerkleCoordinator.persistenceBuffer 字段
- persistTransaction() 使用 buffer.Add() 替代直接写入
- Close() 关闭缓冲区

性能优化：
- 批量写入可提升性能 5-10 倍（减少 IO 次数）
- 最大延迟 100ms（maxInterval）

单元测试覆盖：
- 创建缓冲区（默认/自定义配置）
- 添加事务
- 手动/批量/定时刷新
- 并发安全
- 关闭刷新
- 统计信息

* fix(consistency): 移除同义反复的条件检查

在 NewTwoPCMerkleCoordinatorWithTransport 中：
- Line 292 已验证 metadataKV != nil（否则返回错误）
- Line 317 的 if metadataKV != nil 检查永远为 true

修复：直接初始化 persistenceBuffer，移除冗余的条件检查

* feat(quorum): P1-3 Quorum 网络集成 + 测试覆盖率优化

主要变更：
- 实现 Quorum 网络集成器 (NetworkIntegrator)
- 使用 msgpack 替代 gob 进行消息编解码
- 添加压力测试 (500并发/5000 msg/s/1000 tx/s)

P1/P2 问题修复：
- 修复 WaitAll/Wait 方法的 goroutine 泄漏
- 添加 CI 环境兼容的测试阈值
- 改进 Mock Transport 线程安全
- 添加负面测试用例 (空/nil 参与者)

测试覆盖率：quorum 82.1%, consistency 80.5%

* fix(ci): 跳过 CI 环境中的性能测试

在 CI 环境中跳过 TestIntegration_Performance_DifferenceDetection 测试，
因为 CI 环境负载高且不稳定，性能测试结果不可靠。

与 TestIntegration_Performance_MerkleUpdate 保持一致的策略。

+++ internal/rpc/client.go
@@ -23,7 +23,7 @@ type Client struct {
 	host           host.Host
 	pool           *ConnectionPool
 	defaultTimeout time.Duration
-	maxMessageSize uint16
+	maxMessageSize int
 	codec          *transport.MessagePackCodec
 	enablePool     bool // 是否启用连接池
 }
